src/test/java/net/brabenetz/lib/securedproperties/config/ConfigKeyTest.java
/*-
 * #%L
 * Secured Properties
 * ===============================================================
 * Copyright (C) 2016 - 2019 Brabenetz Harald, Austria
 * ===============================================================
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * 
 *      http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 * #L%
 */
package net.brabenetz.lib.securedproperties.config;

import org.junit.Test;

import static org.hamcrest.CoreMatchers.equalTo;
import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;

public class ConfigKeyTest {

    @Test
    public void testGetKebabCase() throws Exception {
        assertThat(ConfigKey.DEFAULT_PREFIX_KEBAB_CASE, is(equalTo("secured-properties")));
        assertThat(ConfigKey.SECRET_FILE.getKebabCase(), is(equalTo("secret-file")));
    }
}
